Work on foods page

+++ app/actions/foodActions.js
@@ -15,14 +15,14 @@ const pool = mysql.createPool({
     keepAliveInitialDelay: 0
 }).promise();
 
-async function createFood(foodName, calories, date, userId, page) {
+async function createFood(foodName, calories, date, userId, meal, page) {
     const session = await getServerSession();
     if (!session) {
         return false;
     }
-    const insertQuery = "INSERT INTO joshy_fitness_tracker_food (name, calories, date, user_id) VALUES (?, ?, ?, ?)";
+    const insertQuery = "INSERT INTO joshy_fitness_tracker_food (name, calories, date, meal, user_id) VALUES (?, ?, ?, ?, ?)";
     try {
-        const [res] = await pool.query(insertQuery, [foodName, calories, date, userId]);
+        const [res] = await pool.query(insertQuery, [foodName, calories, date, meal, userId]);
         revalidatePath(page);
         return res.affectedRows === 1;
     }
@@ -32,6 +32,10 @@ async function createFood(foodName, calories, date, userId, page) {
 }
 
 async function deleteFood(id, page) {
+    const session = await getServerSession();
+    if (!session) {
+        return false;
+    }
     const deleteQuery = "DELETE FROM joshy_fitness_tracker_food WHERE id = ?";
     try {
         const [res] = await pool.query(deleteQuery, [id]);
@@ -44,6 +48,10 @@ async function deleteFood(id, page) {
 }
 
 async function updateFood(id, foodName, calories, page) {
+    const session = await getServerSession();
+    if (!session) {
+        return false;
+    }
     const updateQuery = "UPDATE joshy_fitness_tracker_food SET name = ?, calories = ? WHERE id = ?";
     try {
         const [res] = await pool.query(updateQuery, [foodName, calories, id]);
@@ -55,8 +63,15 @@ async function updateFood(id, foodName, calories, page) {
     }
 }
 
-async function getFoodsForDay(day, page) {
-
+async function getFoodsForDay(day, userId) {
+    const grabFoodsQuery = "SELECT * FROM joshy_fitness_tracker_food WHERE date = ? AND user_id = ?";
+    try {
+        const [rows] = await pool.query(grabFoodsQuery, [day, userId]);
+        return rows;
+    }
+    catch (err) {
+        return false;
+    } 
 }
 
 export { createFood, deleteFood, updateFood, getFoodsForDay }
